+++ sites/public/pages/applications/financial/vouchers.tsx
@@ -108,6 +108,7 @@ const ApplicationVouchers = () => {
                 register={register}
                 validation={{ required: true }}
                 fields={incomeVouchersValues}
+                dataTestId={"app-income-vouchers"}
               />
             </fieldset>
           </div>
@@ -117,6 +118,7 @@ const ApplicationVouchers = () => {
               <Button
                 styleType={AppearanceStyleType.primary}
                 onClick={() => conductor.setNavigatedBack(false)}
+                data-test-id={"app-next-step-button"}
               >
                 {t("t.next")}
               </Button>