+++ sites/public/pages/applications/household/programs.tsx
@@ -111,9 +111,11 @@ const ApplicationPrograms = () => {
         />
 
         <div className="form-card__lead border-b">
-          <h2 className="form-card__title is-borderless">{pageProgram.description}</h2>
+          <h2 className="form-card__title is-borderless">{pageProgram?.description}</h2>
 
-          {pageProgram.subTitle !== "" && <p className="field-note mt-5">{pageProgram.subtitle}</p>}
+          {pageProgram?.subTitle && pageProgram?.subTitle !== "" && (
+            <p className="field-note mt-5">{pageProgram.subtitle}</p>
+          )}
         </div>
 
         {!!Object.keys(errors).length && (
@@ -139,6 +141,7 @@ const ApplicationPrograms = () => {
                       errorMessage={t("errors.selectAnOption")}
                       register={register}
                       validation={{ required: true }}
+                      dataTestId={"app-program-option"}
                       fields={pageProgram?.formMetadata?.options?.map((option) => {
                         return {
                           id: option.key,
@@ -171,6 +174,7 @@ const ApplicationPrograms = () => {
                     conductor.returnToReview = false
                     conductor.setNavigatedBack(false)
                   }}
+                  data-test-id={"app-next-step-button"}
                 >
                   {t("t.next")}
                 </Button>