+++ sites/public/pages/applications/preferences/general.tsx
@@ -63,6 +63,7 @@ const ApplicationPreferencesGeneral = () => {
                   conductor.returnToReview = false
                   conductor.setNavigatedBack(false)
                 }}
+                data-test-id={"app-next-step-button"}
               >
                 {t("t.next")}
               </Button>