+++ sites/public/pages/applications/review/summary.tsx
@@ -51,7 +51,9 @@ const ApplicationSummary = () => {
         <div className="form-card__pager">
           <div className="form-card__pager-row primary">
             <Form onSubmit={handleSubmit(onSubmit)}>
-              <Button styleType={AppearanceStyleType.primary}>{t("t.confirm")}</Button>
+              <Button styleType={AppearanceStyleType.primary} data-test-id={"app-summary-confirm"}>
+                {t("t.confirm")}
+              </Button>
             </Form>
           </div>
         </div>