+++ sites/public/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.1-alpha.16](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.15...@bloom-housing/public@3.0.1-alpha.16) (2021-11-15)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
+## [3.0.1-alpha.15](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.14...@bloom-housing/public@3.0.1-alpha.15) (2021-11-15)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
+## [3.0.1-alpha.14](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.13...@bloom-housing/public@3.0.1-alpha.14) (2021-11-12)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
+## [3.0.1-alpha.13](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.12...@bloom-housing/public@3.0.1-alpha.13) (2021-11-12)
+
+
+### Features
+
+* refactor and add public site application flow cypress tests ([#2118](https://github.com/bloom-housing/bloom/issues/2118)) ([9ec0e8d](https://github.com/bloom-housing/bloom/commit/9ec0e8d05f9570773110754e7fdaf49254d1eab8))
+
+
+
+
+
+## [3.0.1-alpha.12](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.11...@bloom-housing/public@3.0.1-alpha.12) (2021-11-12)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
+## [3.0.1-alpha.11](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.10...@bloom-housing/public@3.0.1-alpha.11) (2021-11-11)
+
+**Note:** Version bump only for package @bloom-housing/public
+
+
+
+
+
 ## [3.0.1-alpha.10](https://github.com/bloom-housing/bloom/compare/@bloom-housing/public@3.0.1-alpha.9...@bloom-housing/public@3.0.1-alpha.10) (2021-11-10)
 
 **Note:** Version bump only for package @bloom-housing/public